Part 4.15-4.22

+++ controllers/blogs.js
@@ -1,23 +1,38 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (req, res) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({})
+        .populate({ path: 'user', select: ['username', 'name', 'id']})
     await res.json(blogs.map(blog => blog.toJSON()))
 })
 
-blogsRouter.post('/', async (req, res) => {
-    const blog = new Blog({
-        url: req.body.url,
-        title: req.body.title,
-        author: req.body.author,
-        likes: req.body.likes || 0
-    })
+blogsRouter.post('/', async (req, res, next) => {
     try {
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+        const user = await User.findById(decodedToken.id)
+
+        const blog = new Blog({
+            url: req.body.url,
+            title: req.body.title,
+            author: req.body.author,
+            likes: req.body.likes || 0,
+            user: user.id
+        })
         const savedBlog = await blog.save()
-        await res.status(201).json(savedBlog.toJSON)
+
+        // Add blog to the corresponding user
+        user.blogs = user.blogs.concat(savedBlog.id)
+        await user.save()
+
+        await res
+            .status(201)
+            .json(savedBlog.toJSON())
     } catch (exception) {
-        res.status(400).end()
+        next(exception)
     }
 })
 
@@ -42,16 +57,25 @@ blogsRouter.put('/:id', async (req, res) => {
     }
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', async (req, res, next) => {
     try {
-        const deletedBlog = await Blog.findByIdAndRemove(req.params.id)
-        if (deletedBlog) {
-            res.status(204).end()
+        const blog = await Blog.findById(req.params.id)
+
+        // Blog not found
+        if (!blog) {
+            return res.status(404).json({ error: 'blog not found' })
+        }
+
+        const decodedToken = jwt.verify(req.token, process.env.SECRET)
+        // Verify that deletion is done by the blog's creator
+        if (blog.user.toString() === decodedToken.id) {
+            await Blog.findByIdAndRemove(req.params.id)
+            return res.status(204).end()
         } else {
-            res.status(400).end()
+            return res.status(401).json({ error: 'deletion only possible by the creator' })
         }
     } catch (exception) {
-        res.status(400).end()
+        next(exception)
     }
 })
